Add import to transition module
This commit fixes issue #32. This was triggered because paper-dialog had
hard dependency to the paper-dialog-transition component, even though
documentation said it was optional. We fix this by importing transition
component from paper-dialog and updating documentation to reflect the
change.

Documentation now also highlights transition attribute in examples to
make them more visible in component description page, as currently the
values of transition attributes are clipped.